Split CSS file into components scss
every component has its own SCSS files, no matter if it's going to be
kept after this refactor or deleted. Just to have a starting point.

Each SCSS file content was created from CSS using Chat GPT to transform
it, so it might have some errors. Every SCSS file needs to be doubled
checked and this must be done at the moment of refactoring each
component.